Move API implementation to its own file
The API implementation is now in api.c
