+++ core/linux/daemon/httpd.c
@@ -21,25 +21,26 @@ along with Armadito core.  If not, see <http://www.gnu.org/licenses/>.
 
 #define _GNU_SOURCE
 
-#include <sys/types.h>
+/* why first??? #include <sys/types.h> */
 #ifndef _WIN32
 #include <sys/select.h>
 #include <sys/socket.h>
 #else
 #include <winsock2.h>
 #endif
+#include <microhttpd.h>
+
 #include <arpa/inet.h>
-#include <sys/types.h>
 #include <errno.h>
 #include <fcntl.h>
 #include <glib.h>
 #include <json.h>
 #include <magic.h>
-#include <microhttpd.h>
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/types.h>
 #include <sys/stat.h>
 #include <time.h>
 #ifdef _WIN32
@@ -51,13 +52,12 @@ along with Armadito core.  If not, see <http://www.gnu.org/licenses/>.
 #include <libarmadito.h>
 
 #include "httpd.h"
+#include "api.h"
 #include "reportp.h"
 
 #define PAGE_400 "<html><head><title>forbidden!</title></head><body>Bad Request. Make sure your request has a X-Armadito-Token header.</body></html>\n"
 #define PAGE_403 "<html><head><title>forbidden!</title></head><body>Request forbidden. Make sure your request has a User-Agent header.</body></html>\n"
 #define PAGE_404 "<html><head><title>not found!</title></head><body>not found!</body></html>\n"
-
-#define A6O_HTTP_HEADER_TOKEN "X-Armadito-Token"
 
 #undef USE_GLIB_CHANNEL
 #define USE_MHD_THREAD_POOL
@@ -69,7 +69,6 @@ struct httpd {
 	struct MHD_Response *response_400;
 	struct MHD_Response *response_403;
 	struct MHD_Response *response_404;
-	GHashTable *client_table;
 	magic_t magic;
 };
 
@@ -190,346 +189,6 @@ static int content_serve(struct httpd *h, struct MHD_Connection *connection, con
 	 MHD_destroy_response(response);
 
 	 return ret;
-}
-
-#if 0
-/*
-   OnDemandProgressEvent:
-    allOf:
-      - $ref: '#/definitions/Event'
-      - type: object
-        properties:
-          progress:
-            type: integer
-            format: int32
-            description: the progress bar
-          malware_count:
-            type: integer
-            format: int32
-            description: malware counter
-          suspicious_count:
-            type: integer
-            format: int32
-            description: suspicious counter
-          scanned_count:
-            type: integer
-            format: int32
-            description: scanned file counter
-* */
-
-static struct a6o_report *report_fake_new(void)
-{
-	static int count = 0;
-	struct a6o_report *r = malloc(sizeof(struct a6o_report));
-
-	a6o_report_init(r, 42, "/foo/bar", count);
-	count++;
-	count %= 101;
-
-	return r;
-}
-
-static struct json_object *json_on_demand_progress_event_new(struct a6o_report *report)
-{
-	struct json_object *j_report;;
-
-	j_report = json_object_new_object();
-
-	json_object_object_add(j_report, "eventType", json_object_new_string("OnDemandProgressEvent"));
- 	json_object_object_add(j_report, "progress", json_object_new_int(report->progress));
-	json_object_object_add(j_report, "malware_count", json_object_new_int(report->malware_count));
-	json_object_object_add(j_report, "suspicious_count", json_object_new_int(report->suspicious_count));
-	json_object_object_add(j_report, "scanned_count", json_object_new_int(report->scanned_count));
-
-	return j_report;
-}
-#endif
-
-struct api_client {
-	GAsyncQueue *event_queue;
-};
-
-static struct api_client *api_client_new(void)
-{
-	struct api_client *c = malloc(sizeof(struct api_client));
-
-	c->event_queue = g_async_queue_new();
-
-	return c;
-}
-
-static void api_client_destroy(struct api_client *c)
-{
-	g_async_queue_unref(c->event_queue);
-}
-
-typedef int (*api_cb_t)(struct httpd *h, struct MHD_Connection *connection, struct json_object **out);
-
-static int token_api_cb(struct httpd *h, struct MHD_Connection *connection, struct json_object **out);
-static int ping_api_cb(struct httpd *h, struct MHD_Connection *connection, struct json_object **out);
-static int scan_api_cb(struct httpd *h, struct MHD_Connection *connection, struct json_object **out);
-static int poll_api_cb(struct httpd *h, struct MHD_Connection *connection, struct json_object **out);
-
-static struct api_dispatch_entry {
-	const char *path;
-	api_cb_t api_cb;
-} api_dispatch_table[] = {
-	{ "/token", token_api_cb},
-	{ "/ping", ping_api_cb},
-	{ "/scan", scan_api_cb},
-	{ "/poll", poll_api_cb},
-	{ NULL, NULL},
-};
-
-static api_cb_t get_api_cb(const char *path)
-{
-	struct api_dispatch_entry *p;
-
-	for (p = api_dispatch_table; p->path != NULL && strcmp(p->path, path); p++)
-		;
-
-	if (p->path != NULL)
-		return p->api_cb;
-
-	return NULL;
-}
-
-
-static const char *api_get_user_agent(struct MHD_Connection *connection)
-{
-	return MHD_lookup_connection_value(connection, MHD_HEADER_KIND, MHD_HTTP_HEADER_USER_AGENT);
-}
-
-static const char *api_get_token(struct MHD_Connection *connection, int64_t *p_token)
-{
-	const char *s_token;
-
-	s_token = MHD_lookup_connection_value(connection, MHD_HEADER_KIND, A6O_HTTP_HEADER_TOKEN);
-	if (s_token != NULL)
-		*p_token = atoll(s_token);
-
-	return s_token;
-}
-
-#define HASH_ONE(H, C) (H) ^= ((H) << 5) + ((H) >> 2) + (C)
-/* #define HASH_ONE(H, C) (H) = (((H) << 5) + (H)) + (C) */
-
-static void hash_init(int64_t *hash)
-{
-	/* *hash = 5381;*/
-	*hash = 0;
-}
-
-static void hash_buff(const char *buff, size_t len, int64_t *hash)
-{
-	for ( ; len--; buff++)
-		HASH_ONE(*hash, *buff);
-}
-
-static void hash_str(const char *str, int64_t *hash)
-{
-	for ( ; *str; str++)
-		HASH_ONE(*hash, *str);
-}
-
-static int token_api_cb(struct httpd *h, struct MHD_Connection *connection, struct json_object **out)
-{
-	int64_t token;
-	const char *user_agent;
-	char here;
-	time_t now;
-
-	hash_init(&token);
-	time(&now);
-	hash_buff((const char *)&now, sizeof(time_t), &token);
-	user_agent = api_get_user_agent(connection);
-	hash_str(user_agent, &token);
-	hash_buff(&here, sizeof(char *), &token);
-
-	if (token < 0)
-		token = -token;
-
-	a6o_log(ARMADITO_LOG_SERVICE, ARMADITO_LOG_LEVEL_DEBUG, "token %lld", token);
-
-	*out = json_object_new_object();
-	json_object_object_add(*out, "token", json_object_new_int64(token));
-
-	httpd_add_client(h, token);
-
-	return 0;
-}
-
-static int ping_api_cb(struct httpd *h, struct MHD_Connection *connection, struct json_object **out)
-{
-	*out = json_object_new_object();
-	json_object_object_add(*out, "status", json_object_new_string("ok"));
-
-	return 0;
-}
-
-static struct json_object *report_json(struct a6o_report *report)
-{
-	struct json_object *j_report;
-
-	j_report = json_object_new_object();
-
-	json_object_object_add(j_report, "progress", json_object_new_int(report->progress));
-	json_object_object_add(j_report, "malware_count", json_object_new_int(report->malware_count));
-	json_object_object_add(j_report, "suspicious_count", json_object_new_int(report->suspicious_count));
-	json_object_object_add(j_report, "scanned_count", json_object_new_int(report->scanned_count));
-
-	if (report->path != NULL)
-		json_object_object_add(j_report, "path", json_object_new_string(report->path));
-
-	json_object_object_add(j_report, "scan_status", json_object_new_string(a6o_file_status_pretty_str(report->status)));
-	json_object_object_add(j_report, "scan_action", json_object_new_string(a6o_action_pretty_str(report->action)));
-
-	if (report->mod_name != NULL)
-		json_object_object_add(j_report, "mod_name", json_object_new_string(report->mod_name));
-
-	if (report->mod_report != NULL)
-		json_object_object_add(j_report, "mod_report", json_object_new_string(report->mod_report));
-
-	return j_report;
-}
-
-struct scan_data {
-	struct api_client *client;
-	time_t last_send_time;
-	int last_send_progress;
-	struct a6o_on_demand *on_demand;
-};
-
-#ifdef linux
-static time_t get_milliseconds(void)
-{
-	struct timeval now;
-
-	if (gettimeofday(&now, NULL) < 0) {
-		a6o_log(ARMADITO_LOG_MODULE, ARMADITO_LOG_LEVEL_ERROR, "error getting time IHM (%s)", strerror(errno));
-		return 0;
-	}
-
-	return now.tv_sec * 1000 + now.tv_usec / 1000;
-}
-#endif
-
-#ifdef _WIN32
-static time_t get_milliseconds( ) {
-
-	time_t ms = 0;
-	struct _timeb tb;
-
-	_ftime64_s(&tb);
-	ms = tb.time * 1000 + tb.millitm;
-
-	return ms;
-}
-#endif
-
-#define SEND_PERIOD 200  /* milliseconds */
-
-static void scan_callback(struct a6o_report *report, void *callback_data)
-{
-	struct json_object *j_report;
-	struct scan_data *scan_data = (struct scan_data *)callback_data;
-	time_t now = get_milliseconds();
-
-	if (report->status == ARMADITO_CLEAN
-		&& report->progress != 100
-		&& scan_data->last_send_progress != REPORT_PROGRESS_UNKNOWN
-		&& scan_data->last_send_progress == report->progress
-		&& scan_data->last_send_time != 0
-		&& (now - scan_data->last_send_time) < SEND_PERIOD)
-		return;
-
-	j_report = report_json(report);
-
-	g_async_queue_push(scan_data->client->event_queue, j_report);
-
-	scan_data->last_send_time = now;
-	scan_data->last_send_progress = report->progress;
-}
-
-static int scan_api_cb(struct httpd *h, struct MHD_Connection *connection, struct json_object **out)
-{
-	return 0;
-}
-
-static int poll_api_cb(struct httpd *h, struct MHD_Connection *connection, struct json_object **out)
-{
-	int64_t token;
-	struct api_client *client;
-
-	api_get_token(connection, &token);
-	client = httpd_get_client(h, token);
-
-	if (client != NULL) {
-		*out = (struct json_object *)g_async_queue_pop(client->event_queue);
-		return 0;
-	}
-
-	return 1;
-}
-
-static int api_serve(struct httpd *h, struct MHD_Connection *connection, const char *path)
-{
-	const char *json_buff;
-	struct json_object *j_response;
-	api_cb_t api_cb;
-	struct MHD_Response *response;
-	int64_t token;
-	int ret;
-
-	a6o_log(ARMADITO_LOG_SERVICE, ARMADITO_LOG_LEVEL_DEBUG, "request to API: path %s", path);
-
-	/* return a HTTP 404 if path is not valid */
-	api_cb = get_api_cb(path);
-	if (api_cb == NULL) {
-		a6o_log(ARMADITO_LOG_SERVICE, ARMADITO_LOG_LEVEL_WARNING, "request to API invalid path %s", path);
-		return MHD_queue_response(connection, MHD_HTTP_NOT_FOUND, h->response_404);
-	}
-
-	/* return a HTTP 403 (forbidden) if no User-Agent header */
-	if (api_get_user_agent(connection) == NULL) {
-		a6o_log(ARMADITO_LOG_SERVICE, ARMADITO_LOG_LEVEL_WARNING, "request to API path %s has no User-Agent", path);
-		return MHD_queue_response(connection, MHD_HTTP_FORBIDDEN, h->response_403);
-	}
-
-	/* if endpoint is not /token and if no token in HTTP headers, return a HTTP 400 (bad request) */
-	if (strcmp(path, "/token")) {
-		if (api_get_token(connection, &token) == NULL) {
-			a6o_log(ARMADITO_LOG_SERVICE, ARMADITO_LOG_LEVEL_WARNING, "request to API path %s has no token", path);
-			return MHD_queue_response(connection, MHD_HTTP_FORBIDDEN, h->response_400);
-		}
-
-		a6o_log(ARMADITO_LOG_SERVICE, ARMADITO_LOG_LEVEL_DEBUG, "api token %lld", token);
-	}
-
-	/* return a HTTP 400 (bad request) if request parameters are not valid */
-	/* TODO */
-
-	j_response = NULL;
-	ret = (*api_cb)(h, connection, &j_response);
-	json_buff = json_object_to_json_string(j_response);
-
-	response = MHD_create_response_from_buffer(strlen(json_buff), (char *)json_buff, MHD_RESPMEM_MUST_COPY);
-	if (response == NULL) {
-		json_object_put(j_response); /* free the json object */
-		return MHD_NO;
-	}
-
-	json_object_put(j_response); /* free the json object */
-
-	MHD_add_response_header(response, MHD_HTTP_HEADER_CONTENT_TYPE, "application/json");
-	MHD_add_response_header(response, MHD_HTTP_HEADER_CONNECTION, "close");
-	MHD_add_response_header(response, MHD_HTTP_HEADER_ACCESS_CONTROL_ALLOW_ORIGIN, "*");
-	/* Access-Control-Allow-Headers: Origin, X-Requested-With, Content-Type, Accept */
-
-	ret = MHD_queue_response(connection, MHD_HTTP_OK, response);
-	MHD_destroy_response(response);
-
-	return ret;
 }
 
 static const char *get_path(const char *url)
@@ -719,59 +378,6 @@ struct httpd *httpd_new(unsigned short port)
 	return h;
 }
 
-static int httpd_add_client(struct httpd *h, int64_t token)
-{
-	int64_t *p_key;
-	struct api_client *client;
-
-	p_key = malloc(sizeof(int64_t));
-	*p_key = token;
-
-	if (g_hash_table_contains(h->client_table, p_key)) {
-		a6o_log(ARMADITO_LOG_SERVICE, ARMADITO_LOG_LEVEL_WARNING, "API token %lld already registered", token);
-		free(p_key);
-		return 0;
-	}
-
-	client = api_client_new();
-
-	return g_hash_table_insert(h->client_table, p_key, client);
-}
-
-static struct api_client *httpd_get_client(struct httpd *h, int64_t token)
-{
-	int64_t *p_key;
-	struct api_client *c;
-
-	p_key = malloc(sizeof(int64_t));
-	*p_key = token;
-
-	c = g_hash_table_lookup(h->client_table, p_key);
-
-	if (c == NULL)
-		a6o_log(ARMADITO_LOG_SERVICE, ARMADITO_LOG_LEVEL_WARNING, "API token %lld is not registered", token);
-
-	free(p_key);
-
-	return c;
-}
-
-static int httpd_remove_client(struct httpd *h, int64_t token)
-{
-	int64_t *p_key;
-
-	p_key = malloc(sizeof(int64_t));
-	*p_key = token;
-
-	if (!g_hash_table_contains(h->client_table, p_key)) {
-		a6o_log(ARMADITO_LOG_SERVICE, ARMADITO_LOG_LEVEL_WARNING, "API token %lld is not registered", token);
-		free(p_key);
-		return 0;
-	}
-
-	return g_hash_table_remove(h->client_table, p_key);
-}
-
 void httpd_destroy(struct httpd *h)
 {
 	MHD_stop_daemon(h->daemon);
@@ -779,6 +385,8 @@ void httpd_destroy(struct httpd *h)
 
 	g_hash_table_destroy(h->client_table);
 }
+
+
 
 
 /* DEPRECATED CODE */
